veyron/services/mgmt/device/impl,veyron/lib/unixfd: more close-on-exec stuff
Some unrelated changes lumped together in this CL:

In device manager, start using the separate stream manager 'trick' to ensure
that the connection we use to the agent when setting up the principal is
closed.

Also in device manager, remove the 'start lock', now that go/vcl/2052 has
ostensibly rendered it useless: there is no more danger of leaking the agent
fd to unwanted children.

In unixfd, set close on exec on a couple fd's we obtain with Dup.

In modules/shell, close the conn file if we err out before we pass the conn to
the net connection.

Change-Id: I98a0b53425094affa2f3ca372495d1b345b5cc9e