ipc: Add an "AllowedServersPolicy" option.
This option allows one to specify a "policy" on which servers
a client will talk to (authenticate). For example, if an application
wants to trust only servers that match the blessing pattern
"foo/...", irrespective of how the object address for that server
was obtained, the "AllowedServersPolicy" option can be used on
every RPC.

The shorter term motivation for this is to have this option used
when traversing the namespace, where the mounttable will provide
patterns that should be matched by the blessings presented by
the next hop (including the final server).

Change-Id: I8248e472844dc91acf312c0b9f6f4666dc286de0