veyron/services/mgmt/device: Add persistent arguments
With this change, device manager arguments that should be persistent are
saved in a separate file. When an update happens, the persistent
arguments are merged with the arguments from the new envelope.

Also updated test.sh to include updating and reverting the device
manager. Without the persistent arguments, the new tests would fail.

Change-Id: Ib69824f64aa2cec9d9e7f902b5abe1552b8102b9