veyron/{lib/exec,services/mgmt/device}: Get app pid
As a consequence of using go's builtin support for running a process
as a different uid, suidhelper forks prior to exec-ing an app at the
behest of the device manager. Tracking the lifetime of the app process
requires its pid, not the pid of the (short lived) suidhelper
application.

This CL extends the exec facility to permit the app itself to indicate
its pid as part of indicating its readiness.

Change-Id: I433450fd9164b731c1d719f6f225e02e6586e285