veyron/tools/mgmt/device: add install-local command
This allows a user to specify an app using a local binary and env/flags given on
the command-line, instead of having to use a remote application and binary
server. The implementation starts a short-lived app and binary server within the
command-line tool. Currently, this server listens on its own port (or proxy),
but eventually we should share the connection which the tool makes to the device
manager.

Change-Id: Iea0653f37584797ae09a5563ee9e7669f3bcf9e1