security: Reduce usage of ValidatorVOM outside the veyron2/security
This will help with the change to start using CaveatDescriptor
and registration of validation functions (instead of relying on
the package path of types encoded in ValidatorVOM).

General philosophy is to reduce the need of client code to
poke into the details of the Caveat type encoded "on the wire"
(i.e., defined in VDL).

Change-Id: I27a8aa617211adccdda67b2a3c041b9f686acc76