"veyron/runtimes/google/rt": Make VeyronCredentials take precendence over an agent
Going forward we envision all Vandium processes to run under an agent.
The VeyronCredentials environement variable and flag would only be used
when there is a specific use-case that is better handled by a
credentials directory. In such cases, however, we assume that the user explicitly
intends to initialize the principal via the credentials directory, and therefore
we give the credentials directory precedence over the agent.

Thus the order is:
* --veyron.credentials flag takes precedence over
* VEYRON_CREDENTIALS environment variable which takes predence over
* agent

Change-Id: I05dcb17fc5f48260eb259e5ed3c8c7d10a8669a5