veyron/tools/mgmt/device,veyron/services/mgmt/device/impl: config override for local-install
This changelist enables updates for apps installed with local-install. There are
two changes that make this work:

1. allow install-local to specify a config override (like install), so we can
specify configuration for the app independently of what will be present in the
app envelope downloaded during update. We use a command-line flag for this (and
change the "install" command to also use the flag instead of a positional arg).

2. allow install-local to specify an origin for the app (to override the default
origin which points back to the server run by the command-line tool). We use a
setting in the config to convey this information.
This is a convenient way to use for now to get things working. Long-term, we may
drop this in favor of other mechanisms (e.g. UpdateTo)

Change-Id: I363737c5013f510b3cdf6e3544e96f8d13c6ae2a