security/agent: Bugfix for passphrase encoded private keys.
Prior to this change, if you had a passphrase on the private key, then
the agent would fail to start the child process. Likely a fallout from
the veyron2.Init refactorings.

This commit makes the agent usable again with passphrase protected
keys and adds a shell test that would fail without the corresponding
change to agentd/main.go.

Change-Id: I1fd959cc4609b3f6b6a6ac7ab384f8e7be9e5660