veyron/services/mgmt/device: Add local namespace
This change adds a local mounttable and neighborhood service to the
device manager. The applications receive the address of this mounttable
as default namespace root.

The local namespace looks like:
  <root>/devmgr  -> device manager server
  <root>/global  -> the global namespace
  <root>/nh      -> the local neighborhood

Each app gets the local namespace as namespace root, and the local
mounttable is the only thing that gets published directly in the global
namespace.

The global namespace looks like:
  <root>/devices/<hostname>/devmgr -> <hostname>'s device manager
  <root>/devices/<hostname>/global -> points back to <root>
  <root>/devices/<hostname>/nh     -> <hostname>'s local neighborhood

Change-Id: I2b816acd3e3513bd6f82e3b818a6b72cd4d15128