veyron/services/mgmt/device: Clean up / fix ACLs
There are currently 3 different ways to manage ACLs in the device
manager:

1. --veyron.acl.file or --veyron.acl.literal
This is used by the reserved dispatcher for objects under __debug.
Before this change, it was also used as a mechanism to override the
device manager's ACLs. This change removes that last part.

2. veyron/services/mgmt/lib/acls
This is the library that's used by applicationd, binaryd, and deviced to
manage and persist ACLs for the mgmt services.

3. the mounttable's own ACL
This is currently a single file that gets read once when the mounttable
server is started. The ACLs can be updated with the SetACL method, but
changes are not yet persistent.

This change makes it so that all three mechanisms can coexist while we
look for ways to remove redundancy.

Change-Id: Ic8b12f966ca7c9e76c8f503e08efe3a8af926b9f