v.io/core/veyron2/verror2:  First step is making verror2.E be a struct
This CL makes verror2.E still be an interface (that implements "error"),
but removes from it all other methods.
Now, all operations are accessed via standalone functions in verror2.

Operations like Make() and Convert() now return an "error", as opposed to a verror2.E.
The intent is to make "error" be the only interface tyype that clients deal with.

There is a companion CL
	https://vanadium-review.googlesource.com/#/c/3205/
which changes part of wspr to use error instead of verror2.E in a couple of places.

Tested with
	cd release/go/src/v.io;
	v23 go test ./...
	v23 test run vanadium-go-test

Change-Id: I0154f12fab9e11a32e2c5a7374fcc7bfbc5aebe8