veyron/runtimes/google/naming/namespace: Run globAtServer in parallel
With this change, globAtServer() will initiate a StartCall to all
servers in parallel. As soon as one returns successfully, it cancels the
others, and continues with the request.

The previous behavior was to try one server at a time until it found one
that works, which means that globAtServer could take up to N * callTimeout
to complete.

Also, increase the callTimeout to 30 sec to deal with our very slow ipc
Stack.

Change-Id: Idebf889f30d8ee6666f7cb211ea25f9fcb66ade4