veyron/services/mgmt/suidhelper: Pass std FDs to the child
If we don't pass stdin, stdout, and stderr explicitly,
syscall.StartProcess will close them and the child will not receive
them. Then, if the child attempts to write to stdout or stderr, or attempts
to read from stdin, it might interfere with other open connections.

With this change, the standard file descriptors are passed explicitly.

Change-Id: I07cf73e089e781830eaeb07a100091453c413ecc