v.io/core/veyron/runtimes/google: convert to use verror2
This CL converts
	v.io/core/veyron/runtimes/google/ipc/stream/vif/vif.go
	v.io/core/veyron/runtimes/google/naming/namespace/namespace.go
to use verror2 instead of verror1.

In the first, a custom error is wrapped in a standard error code (Aborted)
because some other parts of the code (manager.go?) test for Aborted
as a special case.  We may wish to fix this later.

In the second, a custom error code is used.

In addition, some unused routines are removed from verror1.

Tested:
	cd release/go/src/v.io/core
	v23 go test ./...
	v23 test run vanadium-go-test

Change-Id: Icf2000dbe4f7f1ec18568b041e3996f7119d5760