v2/services/mgmt,v/tools/mgmt/device: install package override (vdl+client)
This change modifies the device Install API to accept an override for packages
to install (on top of what's in the envelope). The motivation is to provide
packages like ACL files, which should not come from the publisher in the app
envelope, but rather be specified by the installer.

Part of this CL are also the client-side changes needed to support this,
specifically in the device command-line tool.

Server-side changes to come in a future CL.

Change-Id: Ic97db5163bc064a8cf56b9d62e68b68eaba6364d