veyron/runtimes/google/ipc: Always allow MethodSignature
Before this change, all requests to MethodSignature would be rejected
because there is no MethodTags in the security context. For now, we
decided to allow all requests, i.e. remove the call to authorize().

Added a TODO to decide if we should hide method signatures that the
caller doesn't have access to.

Change-Id: Ib89f3e2f36f3aa7d29f157b883eb4236ee139dbd