wspr: Treat WSPR as an RPC service.  Part 1 of many.
This series of CLs creates an internal RPC service for javascript apps to use.
In this CL we only create an RPC method for Serve (javascript creating a server
through wspr).  I am doing this to reduce the number of paths between wspr and
javascript to make it easier to implement cross cutting features like
deadlines, cancellation, and vtrace.

MultiPart: 2/2
Change-Id: I080074bb91178cb83c7fb8419236f9367911325d