v23tests: use concrete rather than interface types. MultiPart: 1/3
It's very convenient to be able to embed testing.T and
expect.Session in the v23tests.T type. We were doing this
using interfaces which was attractive because we envisaged being
able to substitute different implementations at some point in
the future. However, embedding in this way results in an autogenerated
stack frame which shows up as <autogenerated>:<some random line no>
when t.Logf, t.Errorf are called. The alternative of wrapping these
functions (i.e. have v23tests.T wrap testing.T.Errorf) can't be used
because the implementation of testing.T assumes a fixed call depth
of 3. This problem occurs if the concrete type (with the embedded
testing.T, expect.Session) that implements the interface type is
passed to a function that expects the interface as a parameter, hence
using structs instead of interfaces allows for embedding and correct
file, line number recording. We opt to lose the abstraction and
burden the v23tests.T implementor with having to deal with alternate
implementations in the future in order to give the test writer the
convenience of correct line number logging and embedded which leads
to a more concise API.

Change-Id: Icc3227434e61157e1b6208d643ab3535e71b86ae