veyron/services: porting the identity server from shell to Go
This CL also replaces the use of os.Stderr in the principal tool with
os.Stdout as the identity server integration test is using the output
of the principal tool and, as a general rule, depending on error output
is brittle. In addition, our new Go-based integration test framework,
to which we are porting the identity server test, does not play nicely
with error output.

Change-Id: I38eaabea603ce2b58f82ceac105200960b0adca0