veyron/services/mgmt/device/impl: Clean up logs
Set --log_dir on deviced so that the log files are written to the
correct log directory instead of /tmp.

Remove the "tee" hack that forked stdout and stderr. Instead, look for a
special env variable to decide if stdout and stderr should be redirected
or not. In general, they should always be redirected to files. The only
exception is for tests.

Change-Id: I84d7c7ffb36e583e343016e1b6be55b148c26204