shell/lib,veyron/tools/vrun: have vrun's test.sh use a test mounttable
Currently, vrun does not invoke setup_server_test (because it doesn't want the
credentials management that comes with that), but as a side-effect, it missed
out on the test mounttable (so it ends up using dev.v.io:8101).

This change refactors the mounttable setup part in shell_test.sh so it can be
invoked independently, and changes vrun/test.sh to use it (thereby breaking the
dependency on dev.v.io).

Change-Id: Iacd5672cfe1c746d1c97f6072d8a5dfaed639bf3