veyron/lib/expect: update ExpectSet(Eventually)?RE
This updates ExpectSet(Eventually)?RE to return the matches as soon as
they are seen, rather than after EOF or a timeout occurs.

All unmatched regular expressions are tested against each line as the
line is read. As soon as there are no more unmatched regular
expressions, the matches are returned and no more output is consumed.

Change-Id: If4fc823a9fb47ae47a1968835726d2e68cd1045e