v/services/mgmt/device: move creation of devmgr info from dispatcher to starter
We no longer create the device manager service dispatcher at the start of the
server (if it's unclaimed).  Hence, move the writing of the info file out of the
NewDispatcher factory into starter.Start.  At the same time, remove the device
manager object name from the info and just leave the PID there (and use the PID
for stopping).  This is going to get cleaned up & expanded in Arup's CL
(https://vanadium-review.googlesource.com/#/c/4534), so for now do the bare
minimum to get devicex stop working again pre-claim.

Change-Id: I44c9b33c28de7912600bbeb2bfef2fdf70333087