veyron/services/mgmt/device/starter: cancel context after claiming
There's some issue with client connections to the local mounttable: if we
publish the claimable server on the local mountable, then the next time we
restart the mounttable and it comes back on the same port, subsequent publish
attempts will fail.  My guess is that VC/VIF caching somehow messes up and
forces the client to try and reuse the old connection.

Change-Id: I374a0396ebe83e6e58fbb716198cbfe81a4a9543