TBR: "core/veyron/lib/shell": Setting child principal
Currently the shell sets up the principal for all child processes
it starts and blesses it with the principal that the shell was
started with. This logic however is too complicated and in some cases
leads to the child principal getting multiple copies of the same
blessings.

This CL simplifies the logic so that a child principal only receives
blessings derived from the default blessings of the shell's principal.
These blessings are marked as default and shareable with everyone.

The shell's principal is determined as follows.
sh := NewShell(ctx, p)
If p != nil then sh.principal is p otherwise sh.principal is
veyron2.GetPrincipal(ctx).

MultiPart: 1/2

Change-Id: Ia72a05895892e7ce17111afb30ce3f3d1b4db404