TBR: veyron/services/identity/identityd: start the veyron runtime early on.
Even though we don't need a veyron context until we create an IPC Server, we
need to initialize the veyron runtime at the outset, in order to make identityd
'device manager friendly'.  Specifically, initializing the runtime ensures flag
overrides passed via the vexec mechanism get applied, and the app calls back to
the device manager to let it know it's up.

Change-Id: I4803258de5596794cb4bf82e7703936855bc2a46