veyron/services/mgmt/device/impl: unset NAMESPACE_ROOT in test
If the test is run from an environemnt with NAMESPACE_ROOT set (even if set to
""), the test device manager started by the SelfInstall command will no longer
use the test mounttable.  To limit the effect of the environment within the
test, just unset the NAMESPACE_ROOT env var if it happens to be set.

Change-Id: I7ec9913ced398cd296cbc72739beea6f1713a12e