runtimes/google/ipc/stream: Marshal discharges safely.
This addresses https://github.com/veyron/release-issues/issues/1144.
Without this change, any package renames would break VC authentication
as the discharges being encoded/decoded on the wire were the native
types (not the VDL-defined WireDischarge) and decoding of those
(since Discharge is an interface) is dependent on the fully
qualified type name.

Change-Id: I8fae41b1eb462a9bcf10443284dd68ba282b8f8e