TBR veyron/runtimes/google/ipc/stream/fix: fix a bug in vif cache
Currently, vifs are cached with the actual protocol not with a
client specified protocol. E.g., vif with "wsh" are cached with
"tcp" or "ws". So we cannot find a vif with "wsh" causing unncessary
vif creation. There is also a similar problem with v4 and v6 protocols.
This CL addresses this issue.

MultiPart: 1/2
Change-Id: Id37854ff1c7b515e19bb377b1b9d1e9cc130f178