veyron/services/mgmt: Use meaningful visible names for apps, rather than having them all called "bin" in ps.
Apps now have a name of the form <title>@<binary>/app constructed from the
envelope, such as
     rpsbot@binaryd/rpsbot/x86/20150219/app

The name is passed in argv[0] and always has the suffix /app, ensuring that
log file names will always be named app.INFO, etc as log file names are
constructed from basename(argv[0]). Also, any unprintable
characters in argv[0] are flattened to '_'

For now, the executable on disk continues to be named "bin".

Change-Id: Ie3120310d5f0bb0bbd8a48baa8c6e5971f3d14ef