veyron/tools/mgmt/device/devicex: refactor sudo usage and only use when needed
This change factors out the chmod and setuid logic into functions, and also adds
a run function that, depending on whether the user running the script is the
devicemanager user, decides to run using sudo -u or not.

While at it, also fix suid_test.sh for linux in multi-user mode (ps truncates
the username to 8 characters).

Change-Id: I7adf7aa689fd8d9479086ff17da9026c83ead29c