Remove FindUnusedPort from lib/testutil and tests.
Identityd uses the go http library, in particular
the TLS http server which does not export the net listener
that it creates. We could duplicate the TLS setup code
and then use http.Serve(ln *net.Listener) but that seems
icky, so instead we move FindUnusedPort into the identityd
server itself. This both removes the temptation for others
to use it and reduces the raciness inherent in its use. That is,
it's far less likely that the port will be stolen in the time
it takes the identityd server to call a function or two than
it is when the port is allocated by one process and then passed
to another, especially on heavily loaded jenkins slaves.

Change-Id: I30fa699261d93cb58e20a7c4eff04b1ebd802daf