profiles/internal/rt: re-enable logging of buildinfo in a restricted way
To avoid the stderr spam or the creation of useless log files with just the
buildinfo in them, condition the logging of buildinfo on the user having
configured an explicit log destination (which we take as a hint that they're
serious about collecting logs).  This is a temporary hack until we get the
buildinfo into the log file headers.

Change-Id: I141945bc66bf61a409d7c38e429f0052083863c5