profiles/internal/ipc: Send "Not implemented" error for objects that don't implement Glob
Whenever an object that doesn't itself implement Glob is returned as a
glob result, also send an error to inform the caller that the object
doesn't implement glob.

This means that the caller will now be able to identify nodes that
never have children.

See https://github.com/veyron/release-issues/issues/1021

MultiPart: 1/4
Change-Id: I4a7dce7d193fb87b4d775335cc8e4199588cc0a1