v.io/v.io/x/ref/security/audit: Fix test to account for use of verror in v.io/v23/security
This is the second part of a multipart change that uses verror in v.io/v23/security
See
	https://vanadium-review.googlesource.com/6961

The only change here is that principal_test.go used reflect.DeepEqual on an error,
and this always found taht two errors were distinct because they had different
stack traces.  Alas, DeepEqual is prone to false alarms since it compared implementation
details.

On Asim's advice, I replacec the DeepEqual with a routine that does an error-specific
check if the arguments are arays of errors.

MultiPart: 2/2
Change-Id: I02c2b6989230adb00e71f7c81ba6517de57187d2