ipc/default_authorizer_test: Change use of v23.Init to testContext().
Test files in package ipc should be using testContext().
The reason this was working before was somewhat lucky:
Files in package ipc_test were importing a profile which made
this automatically.
I am working on a way for this packages tests to be cleaner so
we don't have these weird profile issues.

Change-Id: I2b19d8180b678367f3e3462b9ac856d16aadd9df