profiles/internal/ipc: Give GlobNotImplemented its own error ID
This change gives GlobNotImplemented its own error ID, and only return
it when the object that doesn't implement Glob is queried with a non
empty pattern.

The previous behavior was inconsistent when calling Glob directly on the
server vs running it through the namespace with an unresolved name.

If a/b is a mounted server that doesn't implement Glob:
ns.Glob("a/b") would return no match and no error, but
ns.Glob("/b-endpoint" would return a GlobNotImplemented error.

This is because the namespace library avoids talking to the mounted
server.

The new bahavior makes everything consistent.

MultiPart: 2/3
Change-Id: Ice7feceb59cced75d356a2e0f6fef3a348f1bf60