services/identity/server: fix shutdown race
When the shutdown sequence begins, we should first stop the ipc server before
proceeding to call the shutdown method that we got back from
v23.Init. Otherwise, the runtime shutdown may shut down the client used to talk
to the security agent before the server had a chance to unmount itself or
complete its inflight rpcs.

Change-Id: I73618338e888176e745986594ec95575806f013e