mgmt,identity: Use a security setup more representative of what will happen in practice in the integration tests.
Specifically:
(a) Clients and servers run as different principals
(b) Avoid use of the VEYRON_CREDENTIALS environment variable,
    (and utilize the implicit agent in the v23tests framework)

The most notable change is in the DeviceManager test (renamed from
NodeManager in this change). An unclaimed device manager now starts
with self-signed blessings and gains useful ones only when being
"claimed".

Change-Id: I20b901392dd013d0e3aa6f5c6c840656f426b71b