services/mgmt/lib/acls: restructure the hierarchical authorizer for use by debug
The debug implementation uses the same authorizer over the lifetime of
the service. Modify the hierarchical authorizer to support monitoing
permission lists dynamically instead of requiring the creation of a
new Authorizer in each Lookup call.

Re-land with a modification: now makes the same number of attempts to
access the actual Permissions storage file as the upstream code. As a
result, this fixes a timing error in the Darwin integration tests.

Change-Id: Ib9defaa67554f95f7e64404f11fe36d4d9e39727