ref/lib/flags, ref/test: make it possible to set new default values.
Ideally it should be possible for test.Init to set 'safe' defaults
for use in tests - e.g. not to use ns.dev.v.io etc. This is a little
more complicated because of the initialization order between profiles,
the runtime and flags. The flag sets must be created by an init
function as per usual, however, this is also when default values
are read (this is so that --help can display the correct defaults).
This change allows for defaults (via the SetDefaults functions)
to be changed outside of init functions but before flags.Parse is
called.

Fix up the mounttable cmd test to not rely on ns.dev.v.io.

Change-Id: Icc166fff04bb5a263981231a0ace064d5beacc74