profiles/internal: Add IsLeaf
When a server is using LeafDispatcher, it now publishes itself with a
flag that indicates that it is a leaf. This information can be used by
the namespace clients to avoid looking for objects beyond the root of
the server.

GlobResult now also includes the IsLeaf bit as well to indicate which
internal objects are leaves.

BUG=https://github.com/veyron/release-issues/issues/1495

MultiPart: 2/2
Change-Id: Idbae6ad68d19da1d41df7da68d39167e45dcfa10