services/mgmt/device/impl: implement Application.Status server-side
This is goes with the vdl change in go/vcl/8742

Status is implemented for installations and instances. Most of the changes in
this CL are related to the state enum moving into VDL; the Status methods
themselves are small.

While at it, fix logic in the reaper test to:

- remove a theoretical race between app process exiting and device manager
  starting up (unlikely to ever occur in practice)

- correctly update instance states (using initializeState actually adds a second
  state in addition to whatever state was already there).

MultiPart: 2/2

Change-Id: Id0b82a567fb8cc94027a2f3f4112962b52fcd826