v.io/x/ref/security: Use verror instead of fmt.Errorf and errors.New
This changed uses verror instead of fmt.Errorf and errors.New
in all non-test code in v.io/x/ref/security.

Local variables are used for the error codess except for the
error code ErrBadPassphrase in util.go, which corresponds to the
error (as opposed to error code) PassphraseErr in the old code.
The new variable is exported only because the old one was.

In most cases the error codesare defined in the single file where
they are used.  In a few cases, such as ErrBadPassphrase and
errDataOrSignerUnspecified in blessingstore.go, the error code is
also used in one or more other modules.

In all cases, the error message is kept as much as possible as it
was.  The only consistent addition is the "{1:}{2:} " at the
front to output the programme and operation (RPC) names.

Tested:
	cd v.io
	v23 go test ./...
	v23 test run vanadium-go-test

Change-Id: Ic1809c4f6a332f039838cd76c681be9f3a72b4ed