services/identity: Add --mountprefix flag
Add a flag to specify the prefix for the name that identityd published
in the namespace. The default value is "identity", just like the
previously hard-coded value.

I intend to set --mountprefix=/ns.dev.v.io:8101/identity on the
production instance such that the published name does not depend on the
local namespace provided by the device manager.

Change-Id: Icc82fb655fc0947ab406808f5148982864913112