x/ref: Add vdl "file doc", used to insert the copyright header.
This fixes issue:
https://github.com/veyron/release-issues/issues/1599

Before this change, the vdl tool would automatically include the
vanadium copyright header in all generated files.  This means
that users would also get the copyright header in their generated
files, which is wrong.

After this change, we introduce the notion of "file doc" in each
*.vdl file, which is the comment block that starts on line 1 of
the file, as long as it isn't the package doc.  We require that
all files in a vdl package have the same file doc, otherwise we
return a compilation error.  This makes it easier for java and
javascript, which generate files that don't match 1-to-1 with the
source *.vdl files.

The code generators have been updated to use this file doc when
generating their respective files.

Change-Id: Id53e1a37012d6ff29718753aa0ceb947c9ac43d8