services/identity: Adopt privacy team recommendation.
The identity server issues blessings with the email address in it.
Thus any peer (servers, clients) that this blessing is shared with
know the email address of the owner. Email addresses are considered
sensitive in terms of privacy.

Make a note of that on the page that shows the blessing that will be
created (and which will typically include the email address).

I'm not fully satisfied with the text or styling, but hopefully
someone will have a better suggestion :)

Change-Id: Ib0accc27324f46adbb4ffcf37fbde94213baa7a5