services/identity: Fix for the failing production services tests.
A recent change to the production setup causes identityd to issue
blessings of the form dev.v.io/root/... while some services like
applicationd and binaryd run with blessings of the form
dev.v.io/roles/... The root certificate in both cases (dev.v.io) has the
same public key.

The chrome extension and the prod-services-test bootstrap the set of
trusted roots via an HTTP GET to identityd. This GET does provide the
root certificate's public key, but the blessing name of identityd
(dev.v.io/root). Which means that the HTTP handler wouldn't help others
recognize the dev.v.io/role/...  blessings.

For now, we change this handler to provide both the name and public key
of the root certificate.

Change-Id: I4e6868268fc110a5198d5b9f6621d99a0cf1b707