x/ref: Move services/mounttable/lib to mounttablelib
This is to obey our conventions used in every other package under
services.  It's a longer name, but it's not used that often.

In the future we'll figure out whether all of the stuff exported
from the library really need to be; if not we can move some stuf
under an internal directory.

Change-Id: I15cdca43d443a8427928e898fa347c9e010b487e
MultiPart: 1/3