services/mgmt/{device,suidhelper}: Quickly return forked pid from helper
Return the pid of forked child from suidhelper to the device manager, and
avoid waiting the entire timeout if the app dies while the device manager
is waiting for the init handshake, or for the app cycle mgr name.

If the app appears to be lying about its pid, we don't currently do anything
beyond treating this situation as a failure to start the app. We'll make
suidhelper kill the app in a future update.

Change-Id: I53a0fa04ea95f1332ff30bd5131194b44c4b4c32