ref: Use the runtime authorizer for vtrace.
Previously I used the authorizer returned by the users dispatcher.  The problem
is I'm checking for Debug method tag access and there is no guarantee that
the users authorizer knows or cares about Debug.

The reset of the debug namespace uses the runtime authorizer to authorize debug
requests, so I will follow suit with vtrace.

In the process I also rewrote the vtrace tests substatially.

Change-Id: I825aaeb67be4fe12a4cf47d9d11e468cac4c7293