ref/v23tests: improve concurrent access to V23_BIN_DIR
Prior to this CL, multiple processes sharing a V23_BIN_DIR would
concurrently build and copy over the same binaries. This CL
ensures that every build uses a different output name and then
uses the atomic Rename operation to move the new binary to its
final name. Whilst working on this it seems that concurrent Renames
will not be serialized, rather all-but-one, will fail, hence if the
backoff code that retries the rename after a random period.

Change-Id: Ie248c3134574b36ca8241d12076b9073d4926057